requirements.txt
streamlit==1.30.0
google-generativeai==0.4.0
requests==2.31.0
pandas==2.2.0
plotly==5.20.0
python-dotenv==1.0.1
